add status to initiative file
