+++ app/action-endpoint.ts
@@ -1,13 +1,12 @@
 import { DynamoDB } from 'aws-sdk';
 import { apiWrapper, ApiSignature } from '@manwaring/lambda-wrapper';
+import { Message, Payload } from 'slack';
 import { InitiativeAction, MemberAction, StatusUpdateAction } from './interactions';
 import { CreateMemberRequest, MEMBER_TYPE, MemberResponse, DeleteMemberRequest } from './member';
-import { INITIATIVE_TYPE, InitiativeRecord, InitiativeResponse } from './initiative';
+import { INITIATIVE_TYPE, InitiativeRecord, InitiativeResponse, Status } from './initiative';
 import { DetailResponse } from './slack-responses/detail-response';
 import { getUserProfile } from './slack-calls/profile';
 import { NotImplementedResponse } from './slack-responses/not-implemented-response';
-import { Message, Payload } from 'slack';
-import { Status } from './status';
 import { send } from './slack-calls/send-message';
 
 const initiatives = new DynamoDB.DocumentClient({ region: process.env.REGION });
@@ -36,14 +35,6 @@ export const handler = apiWrapper(async ({ body, success, error }: ApiSignature)
         response = new DetailResponse(initiative, slackUserId, channel);
         break;
       }
-      // case MemberAction.MAKE_MEMBER:
-      // case MemberAction.MAKE_CHAMPION: {
-      //   const { initiativeId, slackUserId, champion } = JSON.parse(payload.actions[0].value);
-      //   await changeMembership(initiativeId, slackUserId, champion);
-      //   const initiative = await getInitiativeDetails(initiativeId);
-      //   response = new DetailResponse(initiative, slackUserId, channel);
-      //   break;
-      // }
       case MemberAction.MAKE_CHAMPION: {
         const { initiativeId, slackUserId } = JSON.parse(payload.actions[0].value);
         await changeMembership(initiativeId, slackUserId, true);
@@ -65,14 +56,6 @@ export const handler = apiWrapper(async ({ body, success, error }: ApiSignature)
         response = new DetailResponse(initiative, slackUserId, channel);
         break;
       }
-      // case InitiativeAction.UPDATE_STATUS: {
-      //   const { initiativeId, status } = JSON.parse(payload.actions[0].selected_option.value);
-      //   const slackUserId = payload.user.id;
-      //   await updateInitiativeStatus(initiativeId, status);
-      //   const initiative = await getInitiativeDetails(initiativeId);
-      //   response = new DetailResponse(initiative, slackUserId, channel);
-      //   break;
-      // }
       case InitiativeAction.UPDATE_STATUS:
       case StatusUpdateAction.MARK_ON_HOLD:
       case StatusUpdateAction.MARK_ABANDONED:
