+++ app/list-initiatives.ts
@@ -1,8 +1,7 @@
 import { DynamoDB } from 'aws-sdk';
 import { apiWrapper, ApiSignature } from '@manwaring/lambda-wrapper';
-import { INITIATIVE_TYPE, InitiativeRecord, InitiativeResponse } from './initiative';
+import { INITIATIVE_TYPE, InitiativeRecord, InitiativeResponse, Status } from './initiative';
 import { ListResponse } from './slack-responses/list-response';
-import { Status } from './status';
 
 const initiatives = new DynamoDB.DocumentClient({ region: process.env.REGION });
