--- app/status.ts
@@ -1,6 +1,0 @@
-export enum Status {
-  ACTIVE = 'ACTIVE',
-  COMPLETE = 'COMPLETE',
-  ABANDONED = 'ABANDONED',
-  ON_HOLD = 'ON_HOLD'
-}
